Add filter for desired number of players
There were filters for minimum and maximum player counts, but if there
is group of N players ready to play, there isn't a filter to return
games that can accommodate that group, regardless of min/max player
count.

Add such a filter to select the number of players, which can be set on
its own or in addition to min/max player count filter.